+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Table-v2: fixed save moved column position
+- Table-v2: events do not work when creating a dashboard
 
 ## [1.136.0]